MBS-11502 refactor UploadCdBuildResultTask; add inputs
