+++ subprojects/gradle/artifactory-app-backup/src/main/kotlin/com/avito/android/plugin/artifactory/ArtifactoryAppBackupInterface.kt
@@ -2,7 +2,7 @@
 
 package com.avito.android.plugin.artifactory
 
-import com.avito.cd.BuildVariant
+import com.avito.cd.model.BuildVariant
 import com.avito.kotlin.dsl.ProjectProperty
 import com.avito.kotlin.dsl.getMandatoryStringProperty
 import org.gradle.api.Project
