+++ subprojects/gradle/artifactory-app-backup/src/main/kotlin/com/avito/android/plugin/artifactory/MavenArtifactsToCdBuildResultArtifactMapper.kt
@@ -1,7 +1,7 @@
 package com.avito.android.plugin.artifactory
 
-import com.avito.cd.BuildVariant
 import com.avito.cd.CdBuildResult
+import com.avito.cd.model.BuildVariant
 import org.gradle.api.publish.maven.MavenArtifact
 import java.net.URI
