+++ subprojects/gradle/artifactory-app-backup/src/test/kotlin/com/avito/android/plugin/artifactory/MavenArtifactsToCdBuildResultArtifactMapperTest.kt
@@ -1,14 +1,14 @@
 package com.avito.android.plugin.artifactory
 
-import com.avito.cd.BuildVariant
 import com.avito.cd.CdBuildResult
+import com.avito.cd.model.BuildVariant
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import java.io.File
 import java.net.URI
 
-class MavenArtifactsToCdBuildResultArtifactMapperTest {
+internal class MavenArtifactsToCdBuildResultArtifactMapperTest {
 
     @Test
     fun `maven artifacts to cd build result mapper - success apk mapping`(@TempDir dir: File) {
