+++ subprojects/gradle/cd/src/main/kotlin/com/avito/ci/steps/ArtifactsConfiguration.kt
@@ -1,6 +1,6 @@
 package com.avito.ci.steps
 
-import com.avito.cd.BuildVariant
+import com.avito.cd.model.BuildVariant
 import groovy.lang.Closure
 import org.gradle.api.Action
