+++ subprojects/gradle/upload-cd-build-result/src/main/kotlin/com/avito/cd/UploadCdBuildResultApi.kt
@@ -1,0 +1,3 @@
+package com.avito.cd
+
+public const val uploadCdBuildResultTaskName: String = "uploadCdBuildResult"
