+++ subprojects/gradle/upload-cd-build-result/src/main/kotlin/com/avito/cd/UploadCdBuildResultTask.kt
@@ -1,35 +1,28 @@
 package com.avito.cd
 
 import com.avito.android.androidAppExtension
-import com.avito.git.GitState
 import com.avito.git.gitState
 import com.avito.http.HttpLogger
 import com.avito.logger.GradleLoggerFactory
 import com.avito.utils.gradle.envArgs
-import com.google.gson.Gson
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.Request
-import okhttp3.RequestBody.Companion.toRequestBody
 import org.gradle.api.DefaultTask
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
-import javax.inject.Inject
 
-public const val uploadCdBuildResultTaskName: String = "uploadCdBuildResult"
+public abstract class UploadCdBuildResultTask : DefaultTask() {
 
-public abstract class UploadCdBuildResultTask @Inject constructor(
-    private val uiTestConfiguration: String,
-    private val user: String,
-    private val password: String,
-    private val output: CdBuildConfig.OutputDescriptor,
-    private val suppressErrors: Boolean
-) : DefaultTask() {
+    @get:Input
+    public abstract val uiTestConfiguration: Property<String>
 
-    init {
-        onlyIf {
-            !output.skipUpload
-        }
-    }
+    @get:Input
+    public abstract val user: Property<String>
+
+    @get:Input
+    public abstract val password: Property<String>
+
+    @get:Input
+    public abstract val suppressErrors: Property<Boolean>
 
     private val uploadAction by lazy {
         val logger = GradleLoggerFactory.getLogger(this)
@@ -37,11 +30,11 @@ public abstract class UploadCdBuildResultTask @Inject constructor(
         UploadCdBuildResultTaskAction(
             gson = uploadCdGson,
             client = Providers.client(
-                user = user,
-                password = password,
+                user = user.get(),
+                password = password.get(),
                 logger = HttpLogger(logger)
             ),
-            suppressErrors = suppressErrors
+            suppressErrors = suppressErrors.get()
         )
     }
 
@@ -54,49 +47,7 @@ public abstract class UploadCdBuildResultTask @Inject constructor(
             versionCode = project.androidAppExtension.defaultConfig.versionCode.toString(),
             teamcityUrl = project.envArgs.build.url,
             gitState = gitState.get(),
-            uiTestConfiguration = uiTestConfiguration
+            uiTestConfiguration = uiTestConfiguration.get()
         )
-    }
-}
-
-internal class UploadCdBuildResultTaskAction(
-    private val client: OkHttpClient,
-    private val gson: Gson,
-    private val suppressErrors: Boolean
-) {
-    fun send(
-        buildOutput: BuildOutput,
-        cdBuildConfig: CdBuildConfig,
-        versionCode: String,
-        teamcityUrl: String,
-        gitState: GitState,
-        uiTestConfiguration: String
-    ) {
-        val testResults = buildOutput.testResults[uiTestConfiguration]
-        requireNotNull(testResults) {
-            "Need $uiTestConfiguration testResults on buildOutput"
-        }
-        val result = CdBuildResult(
-            schemaVersion = cdBuildConfig.schemaVersion,
-            buildNumber = versionCode,
-            releaseVersion = cdBuildConfig.releaseVersion,
-            teamcityBuildUrl = teamcityUrl,
-            gitBranch = CdBuildResult.GitBranch(
-                name = gitState.currentBranch.name,
-                commitHash = gitState.currentBranch.commit
-            ),
-            testResults = testResults,
-            artifacts = buildOutput.artifacts
-        )
-        val cdBuildResultRaw = gson.toJson(result)
-        val request = Request.Builder()
-            .url(cdBuildConfig.outputDescriptor.path)
-            .put(cdBuildResultRaw.toRequestBody("application/json".toMediaType()))
-            .build()
-
-        val response = client.newCall(request).execute()
-        if (!suppressErrors && !response.isSuccessful) {
-            throw RuntimeException("Upload build result failed: ${response.code} ${response.body}")
-        }
     }
 }
