+++ subprojects/gradle/upload-cd-build-result/src/main/kotlin/com/avito/cd/UploadCdBuildResultTaskAction.kt
@@ -1,0 +1,50 @@
+package com.avito.cd
+
+import com.avito.git.GitState
+import com.google.gson.Gson
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.OkHttpClient
+import okhttp3.Request
+import okhttp3.RequestBody.Companion.toRequestBody
+
+internal class UploadCdBuildResultTaskAction(
+    private val client: OkHttpClient,
+    private val gson: Gson,
+    private val suppressErrors: Boolean
+) {
+    fun send(
+        buildOutput: BuildOutput,
+        cdBuildConfig: CdBuildConfig,
+        versionCode: String,
+        teamcityUrl: String,
+        gitState: GitState,
+        uiTestConfiguration: String
+    ) {
+        val testResults = buildOutput.testResults[uiTestConfiguration]
+        requireNotNull(testResults) {
+            "Need $uiTestConfiguration testResults on buildOutput"
+        }
+        val result = CdBuildResult(
+            schemaVersion = cdBuildConfig.schemaVersion,
+            buildNumber = versionCode,
+            releaseVersion = cdBuildConfig.releaseVersion,
+            teamcityBuildUrl = teamcityUrl,
+            gitBranch = CdBuildResult.GitBranch(
+                name = gitState.currentBranch.name,
+                commitHash = gitState.currentBranch.commit
+            ),
+            testResults = testResults,
+            artifacts = buildOutput.artifacts
+        )
+        val cdBuildResultRaw = gson.toJson(result)
+        val request = Request.Builder()
+            .url(cdBuildConfig.outputDescriptor.path)
+            .put(cdBuildResultRaw.toRequestBody("application/json".toMediaType()))
+            .build()
+
+        val response = client.newCall(request).execute()
+        if (!suppressErrors && !response.isSuccessful) {
+            throw RuntimeException("Upload build result failed: ${response.code} ${response.body}")
+        }
+    }
+}
