+++ subprojects/gradle/upload-cd-build-result/src/main/kotlin/com/avito/cd/model/AndroidArtifactType.kt
@@ -1,0 +1,11 @@
+package com.avito.cd.model
+
+import com.google.gson.annotations.SerializedName
+
+public enum class AndroidArtifactType {
+    @SerializedName("bundle")
+    BUNDLE,
+
+    @SerializedName("apk")
+    APK
+}
