+++ subprojects/gradle/upload-cd-build-result/src/main/kotlin/com/avito/cd/model/BuildVariant.kt
@@ -1,0 +1,15 @@
+package com.avito.cd.model
+
+import com.google.gson.annotations.SerializedName
+
+public enum class BuildVariant {
+
+    @SerializedName("release")
+    RELEASE,
+
+    @SerializedName("staging")
+    STAGING,
+
+    @SerializedName("debug")
+    DEBUG
+}
