+++ subprojects/gradle/upload-cd-build-result/src/main/kotlin/com/avito/cd/model/NupokatiProject.kt
@@ -1,0 +1,12 @@
+package com.avito.cd.model
+
+import com.google.gson.annotations.SerializedName
+
+public enum class NupokatiProject(public val id: String) {
+
+    @SerializedName("avito")
+    AVITO("avito"),
+
+    @SerializedName("avito_test")
+    AVITO_TEST("avito_test")
+}
