+++ subprojects/gradle/upload-cd-build-result/src/test/kotlin/com/avito/cd/CdBuildConfigDeserializationTest.kt
@@ -4,12 +4,15 @@ import com.avito.cd.CdBuildConfig.Deployment.GooglePlay
 import com.avito.cd.CdBuildConfig.Deployment.Qapps
 import com.avito.cd.CdBuildConfig.Deployment.Track
 import com.avito.cd.CdBuildConfig.Deployment.Unknown
+import com.avito.cd.model.AndroidArtifactType
+import com.avito.cd.model.BuildVariant
+import com.avito.cd.model.NupokatiProject
 import com.google.common.truth.Truth.assertThat
 import com.google.gson.FieldNamingPolicy
 import com.google.gson.GsonBuilder
 import org.junit.jupiter.api.Test
 
-class CdBuildConfigDeserializationTest {
+internal class CdBuildConfigDeserializationTest {
 
     private val gson = GsonBuilder().run {
         setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
@@ -51,7 +54,7 @@ class CdBuildConfigDeserializationTest {
             ]
         }
         """
-        val deserialized = gson.fromJson<CdBuildConfig>(config, CdBuildConfig::class.java)
+        val deserialized = gson.fromJson(config, CdBuildConfig::class.java)
         assertThat(deserialized).isEqualTo(
             CdBuildConfig(
                 schemaVersion = 1,
