+++ subprojects/gradle/upload-cd-build-result/src/test/kotlin/com/avito/cd/CdBuildConfigValidatorTest.kt
@@ -1,10 +1,13 @@
 package com.avito.cd
 
+import com.avito.cd.model.AndroidArtifactType
+import com.avito.cd.model.BuildVariant
+import com.avito.cd.model.NupokatiProject
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
-class CdBuildConfigValidatorTest {
+internal class CdBuildConfigValidatorTest {
 
     @Test
     fun `success - valid contract last supported version`() {
